+++ src/community/optimal_modularity.c
@@ -156,9 +156,10 @@ int igraph_community_optimal_modularity(const igraph_t *graph,
     IGRAPH_CHECK(igraph_strength(graph, &outdegree, igraph_vss_all(),
                                  IGRAPH_OUT, IGRAPH_LOOPS, weights));
 
-    glp_term_out(GLP_OFF);
+    IGRAPH_GLPK_SETUP();
+
     ip = glp_create_prob();
-    IGRAPH_FINALLY(glp_delete_prob, ip);
+    IGRAPH_FINALLY(igraph_i_glp_delete_prob, ip);
 
     glp_set_obj_dir(ip, GLP_MAX);
     st = glp_add_cols(ip, no_of_variables);