+++ src/scg/scg_optimal_method.c
@@ -48,7 +48,7 @@ int igraph_i_optimal_partition(const igraph_real_t *v, int *gr, int n,
                                igraph_real_t *value) {
 
     int i, non_ties, q, j, l, part_ind, col;
-    igraph_i_scg_indval_t *vs = igraph_Calloc(n, igraph_i_scg_indval_t);
+    igraph_i_scg_indval_t *vs = IGRAPH_CALLOC(n, igraph_i_scg_indval_t);
     igraph_real_t *Cv, temp, sumOfSquares;
     igraph_vector_t ps;
     igraph_matrix_t F;
@@ -174,7 +174,7 @@ int igraph_i_optimal_partition(const igraph_real_t *v, int *gr, int n,
 
     igraph_matrix_destroy(&F);
     igraph_matrix_int_destroy(&Q);
-    igraph_Free(vs);
+    IGRAPH_FREE(vs);
     IGRAPH_FINALLY_CLEAN(3);
 
     if (value) {